+++ packages/@react-native-windows/tester/src/js/examples-win/Button/ButtonExample.windows.js
@@ -217,10 +217,11 @@ exports.examples = [
                 testID="accessibility_props"
                 title="Submit Application"
                 accessibilityLabel="Press to submit your application!"
-                accessibilityRole="Button"
+                accessibilityRole="button"
                 accessibilityHint="Submit Button"
                 accessibilityPosInSet={1}
                 accessibilitySetSize={1}
+                accessibilityValue={{Text: "Submit Application"}}
               />
             );
           }}