+++ vnext/Microsoft.ReactNative/Fabric/Composition/UiaHelpers.h
@@ -29,4 +29,6 @@ void UpdateUiaProperty(
     const std::string &oldValue,
     const std::string &newValue) noexcept;
 
+std::string extractAccessibilityValue(const facebook::react::AccessibilityValue &value) noexcept;
+
 } // namespace winrt::Microsoft::ReactNative::implementation